requirements.txt
streamlit
firebase-admin
reportlab
pandas
imagekitio
Pillow
